feat: SD-card loading + boot menu on the Pico (FatFs), menu now RGB565
Bring the boot menu + SD storage to hardware.

- Vendored FatFs R0.16 (third_party/fatfs) with LFN + code page 437.
- sd_spi.c: minimal SD driver over hardware SPI0 (GP2 SCK / GP3 MOSI /
  GP4 MISO / GP5 CS) — clear of HSTX (12-19) and the WeAct PSRAM CS (GP0),
  works on a stock Pico 2. diskio.c bridges it to FatFs.
- storage_pico.c: implements storage.h via FatFs (same API as the SDL
  folder backend).
- menu now renders RGB565 (the Pico's native format, drawn straight into
  the HDMI back buffer; SDL converts 565->ARGB).
- usbkbd: menu mode — keypresses become menu navigation events instead of
  driving the MSX matrix.
- PicoCalcMSX.c: HDMI comes up first, then (option PICOCALC_SD, default ON)
  mount SD, load the BIOS from system/, run the boot menu (USB keyboard),
  load the chosen slot-1 ROM and machine_init. Falls back to the baked-in
  ROMs if the SD can't be mounted.

Both targets build. SDL menu verified; the Pico SD path is untested pending
the SD wiring. Note: large ROMs in RAM are tight on a no-PSRAM Pico 2.

+++ menu.h
@@ -3,8 +3,8 @@
 
 // Platform-agnostic boot menu: a classic MSX-styled config screen (BIOS font +
 // MSX palette) that lets you pick a ROM for slot 1 / slot 2 and a disk for
-// A: / B:. Renders into a 256x192 ARGB (0x00RRGGBB) framebuffer, so the SDL and
-// Pico frontends present it exactly like an emulator frame.
+// A: / B:. Renders into a 256x192 RGB565 framebuffer — the Pico's native format
+// (drawn straight into the HDMI back buffer); the SDL frontend converts to ARGB.
 
 #include <stdint.h>
 #include <stdbool.h>
@@ -29,7 +29,7 @@ typedef enum {
 void menu_init(const uint8_t *bios, menu_config_t *cfg);
 
 void menu_input(menu_input_t in);
-void menu_render(uint32_t *fb); // 256x192 ARGB
+void menu_render(uint16_t *fb); // 256x192 RGB565
 bool menu_start_requested(void);
 
 #endif // MENU_H
